# Conflicts:
#	.vs/KuteShop/v15/.suo
#	App.Aplication/Utils/Utils.cs
#	App.FakeEntity/bin/Debug/App.Aplication.dll
#	App.FakeEntity/bin/Debug/App.Aplication.pdb
#	App.FakeEntity/bin/Debug/App.Core.dll
#	App.FakeEntity/bin/Debug/App.Core.pdb
#	App.FakeEntity/bin/Debug/App.Domain.dll
#	App.FakeEntity/bin/Debug/App.Domain.pdb
#	App.FakeEntity/bin/Debug/App.FakeEntity.dll
#	App.FakeEntity/bin/Debug/App.FakeEntity.pdb
#	App.FakeEntity/bin/Debug/App.Infra.Data.Repository.dll
#	App.FakeEntity/bin/Debug/App.Infra.Data.Repository.pdb
#	App.FakeEntity/bin/Debug/App.Infra.Data.dll
#	App.FakeEntity/bin/Debug/App.Infra.Data.pdb
#	App.FakeEntity/bin/Debug/App.Service.dll
#	App.FakeEntity/bin/Debug/App.Service.pdb
#	App.FakeEntity/obj/Debug/App.FakeEntity.csprojResolveAssemblyReference.cache
#	App.FakeEntity/obj/Debug/App.FakeEntity.dll
#	App.FakeEntity/obj/Debug/App.FakeEntity.pdb
#	App.Front/Views/Checkout/_Checkout.Cart.cshtml

+++ App.Service/Orders/OrderTotalCalculationService.cs
@@ -14,8 +14,8 @@ namespace App.Service.Orders
                 return subTotal;
             }
 
-            subTotal = (from mul in cart
-                        select mul.Quantity * mul.CustomerEnteredPrice).Sum();
+            subTotal = (decimal) (from mul in cart
+                select mul.Quantity * mul.CustomerEnteredPrice).Sum();
 
             return subTotal;
         }
